initiateNewConversation near complete
web/testscripts nothing.

setup heaps with initnewconvo; the structure for each user is steup with a conversations folder, and any user (even new) can be added to a conversation and the conversation will be correctly added.

they are stored as objects within the conversations folder for each person, and store the next ID for the field (although it goes up by 2's atm, working on a fix & needs final testing to ensure 100% working. but the basic concept does!). this can then be referenced like an array although win an index from 1, not 0.

not a whole lot of actual code as there has been a LOT of testing & refining through here to understand how firebase handles some of this shit, how i decided to store it, and how to retrieve it. oh, and where i need to update vs set data. and alot of testing. like alot of chrome console. but from here; just fix a few small bugs to finalize initalisenewconvos, then from here work on the messages and storing them, and then it's onto messages. sick as.